In the folder named 'Testing_accuracies_on_available_data', we have the previously generated features as 'output.csv' upon which @jbnerd has applies various classification algorithms to calculate accuracies with different train_test_splits.
The results are recorded in the file named 'results.odt'.

Evidently Ensemble method algorithms work much better than other static algorithms.

'feature5_extra.py' contains code for the features that are to be generated in addition to
what was already available in 'feature5.py' and 'hasthag.py' from the old repo that we acquired.
This will complete the feature set of article 5.5.2 . 5.5.1 is still left.